+++ Example/Pods/Target Support Files/AirPlay/AirPlay-dummy.m
@@ -1,5 +1,0 @@
-#import <Foundation/Foundation.h>
-@interface PodsDummy_AirPlay : NSObject
-@end
-@implementation PodsDummy_AirPlay
-@end