+++ Example/Pods/Target Support Files/AirPlay/AirPlay-umbrella.h
@@ -1,6 +1,0 @@
-#import <UIKit/UIKit.h>
-
-
-FOUNDATION_EXPORT double AirPlayVersionNumber;
-FOUNDATION_EXPORT const unsigned char AirPlayVersionString[];
-