+++ Sources/AirPlay.h
@@ -1,0 +1,17 @@
+//
+//  AirPlay.h
+//  AirPlay
+//
+//  Created by eMdOS on 2/10/17.
+//
+//
+
+#import <UIKit/UIKit.h>
+
+//! Project version number for AirPlay.
+FOUNDATION_EXPORT double AirPlayVersionNumber;
+
+//! Project version string for AirPlay.
+FOUNDATION_EXPORT const unsigned char AirPlayVersionString[];
+
+// In this header, you should import all the public headers of your framework using statements like #import <AirPlay/PublicHeader.h>